Refactor code, apply conditionals, add RequestMentoring prompt text
