+++ app/javascript/components/editor/FeedbackPanel/FeedbackDetail.tsx
@@ -1,0 +1,27 @@
+import { GraphicalIcon } from '@/components/common'
+import React from 'react'
+
+export function FeedbackDetail({
+  summary,
+  children,
+}: {
+  summary: string
+  children: React.ReactNode
+}): JSX.Element {
+  return (
+    <details className="c-details feedback">
+      <summary className="--summary select-none">
+        <div className="--summary-inner">
+          <span className="summary-title">{summary}</span>
+          <span className="--closed-icon">
+            <GraphicalIcon icon="chevron-right" />
+          </span>
+          <span className="--open-icon">
+            <GraphicalIcon icon="chevron-down" />
+          </span>
+        </div>
+      </summary>
+      {children}
+    </details>
+  )
+}
