+++ app/javascript/components/editor/FeedbackPanel/index.ts
@@ -1,0 +1,1 @@
+export { FeedbackPanel } from './FeedbackPanel'
